feature/#006-Creating_API_Version
- Created the directories 'v1'.
- Moved all the configurations file to directory 'v1'.
- Refactored the test's files.
- Configurated the routes.
- Installed the versionist gem.
