+++ app/models/v1/social_kind.rb
@@ -1,0 +1,5 @@
+module V1
+  class SocialKind < ApplicationRecord
+
+  end
+end
