--- app/serializers/social_kind_serializer.rb
@@ -1,3 +1,0 @@
-class SocialKindSerializer < ActiveModel::Serializer
-  attributes :id, :name
-end
