+++ app/serializers/v1/social_kind_serializer.rb
@@ -1,0 +1,5 @@
+module V1
+  class SocialKindSerializer < ActiveModel::Serializer
+    attributes :id, :name
+  end
+end
