+++ config/routes.rb
@@ -1,4 +1,6 @@
 Rails.application.routes.draw do
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
-  resources :social_kinds, only: [:index, :show]
+  api_version(:module => "V1", :path => {:value => "v1"}) do
+    resources :social_kinds, only: [:index, :show]
+  end
 end
