+++ spec/factories/social_kinds.rb
@@ -1,5 +1,5 @@
 FactoryGirl.define do
-  factory :social_kinds, class: 'SocialKind' do
+  factory :social_kinds, class: 'V1::SocialKind' do
     name 'ExampleName'
   end
 end
